Generate a Signed Web Bundle
This change removes the Service Worker and Web App Manifest generation
and adds support for generating a Signed Web Bundle.

By switching to a static Web App Manifest we work around a bug in the
plugin we were using which prevented the manifest from being included in
the Web Bundle.

Instructions have been updated based on the latest flag names in Chrome.